Add support for the glslang fork of mcshader-lsp
and use it to fix some possible bugs

+++ shaders/lib/material/ao.glsl
@@ -9,6 +9,6 @@ float16_t gen_tex_ao(float16_t srgb_luma, float16_t avg_srgb_luma) {
 #if DIR_SHADING != 0
 	float16_t dir_shading(f16vec3 w_tex_normal) {
 		const float dir_shading = 0.1 * float(10 - DIR_SHADING);
-		return clamp(dot(f16vec3(abs(w_tex_normal.xz), w_tex_normal.y), f16vec3(dir_shading, 0.5 * dir_shading, 1.0)), 0.25 * dir_shading, 1.0);
+		return clamp(dot(f16vec3(abs(w_tex_normal.xz), w_tex_normal.y), f16vec3(dir_shading, 0.5 * dir_shading, 1.0)), float16_t(0.25 * dir_shading), float16_t(1.0));
 	}
 #endif
